Split custom naming schemes manager from detection manager
