+++ core/__init__.py
@@ -9,6 +9,7 @@ if "bpy" not in locals():
     from . import recorder
     from . import retargeting
     from . import detection_manager
+    from . import custom_schemes_manager
     from . import fbx_patcher
     from . import login
     from . import live_data_manager
@@ -24,6 +25,7 @@ else:
     importlib.reload(recorder)
     importlib.reload(retargeting)
     importlib.reload(detection_manager)
+    importlib.reload(custom_schemes_manager)
     importlib.reload(fbx_patcher)
     importlib.reload(login)
     importlib.reload(live_data_manager)
