+++ panels/retargeting.py
@@ -84,11 +84,11 @@ class RetargetingPanel(ToolPanel, bpy.types.Panel):
 
         row = layout.row(align=True)
         row.label(text='Custom Naming Schemes:')
-        row.operator(detector.SaveCustomBonesRetargeting.bl_idname, text='Save')
 
         subrow = layout.row(align=True)
         row = subrow.row(align=True)
         row.scale_y = 0.9
+        row.operator(detector.SaveCustomBonesRetargeting.bl_idname, text='Save')
         row.operator(detector.ImportCustomBones.bl_idname, text='Import')
         row.operator(detector.ExportCustomBones.bl_idname, text='Export')
         row = subrow.row(align=True)
